+++ R/app_ui.R
@@ -74,6 +74,12 @@ dashboard_sidebar <- function() {dashboardSidebar(sidebarMenu(
 
 
   menuItem("In evidenza", icon = icon("bullseye"),
+    menuSubItem("2020-03-21 Alessandria", tabName = "20200321Alessandria",
+                icon = icon("flag")
+    ),
+    menuSubItem("2020-03-21 Vercelli", tabName = "20200321Vercelli",
+                icon = icon("flag")
+    ),
     menuSubItem("2020-03-20 Novara", tabName = "20200320Novara",
                 icon = icon("flag")
     ),
@@ -211,9 +217,9 @@ dashboard_body <- function() {dashboardBody(
         HTML(
           "<ol>
             <li><strong>Home</strong>: Questa pagina.</li>
-            <li><strong>In evidenza</strong>: Principali considerazioni per la regione Veneto.</li>
+            <li><strong>In evidenza</strong>: Considerazioni su temi specifici locali o regionali.</li>
             <li><strong>Andamento epidemia</strong>: Serie temporali dinamiche e interattive.</li>
-            <li><strong>Indici principali</strong>: modelli e previsioni regionali e nazionali.</li>
+            <li><strong>Indici principali</strong>: Principali indici di interesse locale, regionale o nazionale.</li>
             <li><strong>Segnalazioni</strong>: link per segnalazioni tecniche relative all'applicazione.</li>
             <li><strong>Fonti e informazioni</strong>: Descrizione fonti, licenze d'uso e software utilizzati per lo sviluppo.</li>
             <li><strong>Metriche ultima giornata</strong>: Selezionando una regione (o 'Italia', default) vengono visualizzate le principali metriche dell'ultimo aggiornamento dati, che normalmente avviene alle ore 18 da parte della protezione civile.</li>
@@ -232,6 +238,14 @@ dashboard_body <- function() {dashboardBody(
 
   ),
 
+  tabItem(tabName = "20200321Alessandria",
+          h2("Stime previsive sul totale casi ad Alessandria"),
+          mod_focus_20200320_novara_ui("da_alessandria")
+  ),
+  tabItem(tabName = "20200321Vercelli",
+          h2("Stime previsive sul totale casi a Vercelli"),
+          mod_focus_20200320_novara_ui("da_vercelli")
+  ),
   tabItem(tabName = "20200320Novara",
           h2("Stime previsive sul totale casi a Novara"),
           mod_focus_20200320_novara_ui("da_novara")