+++ NEWS.md
@@ -1,3 +1,6 @@
+* Added focus section for Vercelli and Alessandria
+* Update istruzioni (ita)
+
 # covid19ita 0.15.2
 
 * fix date axes in Novara